From excel
Yeah, OK, I used Excel.  I made a hyperbola, xy=0.09.  Then I got
a few points on it, between x=0.1 and 0.9. Then I jittered them by
adding the cumulative exponential distribution of a random number.

Anyway, the point is that it looks like it might be legit data, except
that it's too random.  But for illustration, that's ok.